pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dental-scraper"
version = "0.1.0"
description = "Dental supplies scraper for Brazilian suppliers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "scrapy>=2.11.0",
    "scrapy-playwright>=0.0.40",
    "playwright>=1.40.0",
    "fake-useragent>=1.4.0",
    "rapidfuzz>=3.6.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "unidecode>=1.3.0",
    "w3lib>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
